Move try-catch-block to new init script which triggers all the other scripts to avoid code duplication; #6
