+++ resources/var/tmp/resources/init.groovy.d/scripts/init030installplugins.groovy
@@ -13,9 +13,6 @@ def keyExists(String key){
 	return true
 }
 
-// Try block to stop Jenkins in case an exception occurs in the script
-try {
-
 // configuration
 def plugins = [
   'mailer-plugin',
@@ -56,12 +53,4 @@ for (def shortName : plugins){
 
 if (updateCenter.isRestartRequiredForCompletion()) {
   jenkins.restart();
-}
-
-// Stop Jenkins in case an exception occurs
-} catch (Exception exception){
-	println("An exception occured during initialization");
-	exception.printStackTrace();
-	println("Init script and Jenkins will be stopped now...");
-	throw new Exception("initialization exception")
 }
