+++ resources/var/tmp/resources/init.groovy.d/scripts/init060configuremailer.groovy
@@ -2,9 +2,6 @@ import jenkins.model.*;
 
 // based on https://github.com/r-hub/rhub-jenkins/blob/master/docker-entrypoint.sh#L87
 
-// Try block to stop Jenkins in case an exception occurs in the script
-try {
-
 def instance = Jenkins.getInstance();
 
 // configure mail server
@@ -19,11 +16,3 @@ mailer.save();
 
 
 instance.save();
-
-// Stop Jenkins in case an exception occurs
-} catch (Exception exception){
-  println("An exception occured during initialization");
-  exception.printStackTrace();
-  println("Init script and Jenkins will be stopped now...");
-  throw new Exception("initialization exception")
-}
