+++ resources/var/tmp/resources/init.groovy.d/scripts/init070configurelocation.groovy
@@ -10,9 +10,6 @@ def getValueFromEtcd(String key){
 	return json.node.value
 }
 
-// Try block to stop Jenkins in case an exception occurs in the script
-try {
-
 def instance = Jenkins.getInstance();
 
 // configure jenkins location
@@ -25,11 +22,3 @@ location.setUrl("https://${fqdn}/jenkins");
 location.save()
 
 instance.save();
-
-// Stop Jenkins in case an exception occurs
-} catch (Exception exception){
-  println("An exception occured during initialization");
-  exception.printStackTrace();
-  println("Init script and Jenkins will be stopped now...");
-  throw new Exception("initialization exception")
-}
