+++ resources/var/tmp/resources/init.groovy.d/scripts/init100configuretheme.groovy
@@ -2,9 +2,6 @@ import jenkins.model.*;
 import net.sf.json.JSONObject;
 import org.kohsuke.stapler.*;
 
-// Try block to stop Jenkins in case an exception occurs in the script
-try {
-
 // copy material theme css file to Jenkins' userContent folder
 def srcTheme = new File('/var/tmp/resources/jenkins-material-theme.css')
 def dstTheme = new File('/var/lib/jenkins/userContent/jenkins-material-theme.css')
@@ -22,11 +19,3 @@ simpleThemeDesc.configure(stapler, json);
 
 simpleThemeDesc.save();
 jenkins.save();
-
-// Stop Jenkins in case an exception occurs
-} catch (Exception exception){
-  println("An exception occured during initialization");
-  exception.printStackTrace();
-  println("Init script and Jenkins will be stopped now...");
-  throw new Exception("initialization exception")
-}
