+++ resources/var/tmp/resources/init.groovy.d/scripts/init110disableCoreUpdateMonitor.groovy
@@ -3,21 +3,10 @@
 
 import jenkins.model.*;
 
-// Try block to stop Jenkins in case an exception occurs in the script
-try {
-
 def coreUpdateMonitorID = 'hudson.model.UpdateCenter$CoreUpdateMonitor';
 
 def coreUpdateMonitor = Jenkins.instance.getAdministrativeMonitor(coreUpdateMonitorID);
 if ( coreUpdateMonitor.isEnabled() ) {
     coreUpdateMonitor.disable(true);
     Jenkins.instance.save();
-}
-
-// Stop Jenkins in case an exception occurs
-} catch (Exception exception){
-  println("An exception occured during initialization");
-  exception.printStackTrace();
-  println("Init script and Jenkins will be stopped now...");
-  throw new Exception("initialization exception")
 }
