+++ resources/var/tmp/resources/init.groovy.d/scripts/init990finish.groovy
@@ -24,16 +24,4 @@ def writeState(String state) {
     writeValueToEtcd('state/jenkins', state);
 }
 
-// Try block to stop Jenkins in case an exception occurs in the script
-try {
-
 writeState('ready');
-
-
-// Stop Jenkins in case an exception occurs
-} catch (Exception exception){
-  println("An exception occured during initialization");
-  exception.printStackTrace();
-  println("Init script and Jenkins will be stopped now...");
-  throw new Exception("initialization exception")
-}
